fix(shifu): satisfy adoption source budget
